add correct player position and demo
