--- Assets/LoadSubScene.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 85f3dec10858056489da8c15b5491633
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
