+++ Assets/Scripts/Scriptable Objects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 076416d7c2c865b4c8a29478809016c7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
