+++ Assets/Scripts/Scriptable Objects/PlayerPosition.asset
@@ -1,0 +1,15 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ee1b5a44ca4b55a41aec453de5d04d19, type: 3}
+  m_Name: PlayerPosition
+  m_EditorClassIdentifier: 
+  initialValue: {x: -15, y: 41}
