+++ Assets/Scripts/Scriptable Objects/PlayerPosition.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3eff63ae5dafcd542a14e2d099e52041
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
