+++ Assets/Scripts/Scriptable Objects/VectorValue.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu]
+public class VectorValue : ScriptableObject
+{
+    public Vector2 initialValue;
+}
