+++ Assets/Scripts/Scriptable Objects/VectorValue.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ee1b5a44ca4b55a41aec453de5d04d19
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
