+++ ProjectSettings/EditorBuildSettings.asset
@@ -8,9 +8,9 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/FirstScene.unity
     guid: fb96cd2587f5dd147a7b009269123f82
-  - enabled: 1
-    path: Assets/Scenes/Area 1.unity
-    guid: 0efea8d77b3730846b450365a4a7c9b3
+  - enabled: 0
+    path: 
+    guid: 00000000000000000000000000000000
   - enabled: 1
     path: Assets/Scenes/Player HUD/Player HUD.unity
     guid: 23fc52fa00282ff41b6369d5e30d075c
@@ -26,4 +26,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/NPC_dialogue.unity
     guid: 3f31e10434d918a4baca7140b5b10ce2
+  - enabled: 1
+    path: Assets/Scenes/Dungeon1/Dungeon1.unity
+    guid: 8e82f2dd25845dc40906500d1aaaf071
   m_configObjects: {}
